+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.7.0
+
+- Outputs each `styleHash` with `IconStyle`, `LineStyle` and `PolyStyle` nodes
+- Allows specifying an `iconUrl` in `options` to overwrite the `href` on each `Icon`
+- Upgrade dependencies
+
 ## 0.6.1
 
 - Fix #4 error with non-string id values